// src/codegen/types.ts
export type Duration = {
  height: number;
} | {
  time: number;
};
export type Uint128 = string;
// ...
export type ActiveThreshold = {
  absolute_count: {
    count: Uint128;
  };
} | {
  percentage: {
    percent: Decimal;
  };
};
export type Decimal = string;
export type StakingInfo = {
  existing: {
    staking_contract_address: string;
  };
} | {
  new: {
    staking_code_id: number;
    unstaking_duration?: Duration | null;
  };
};
export type Logo = {
  url: string;
} | {
  embedded: EmbeddedLogo;
};
export type EmbeddedLogo = {
  svg: Binary;
} | {
  png: Binary;
};
export interface Cw20Coin {
  address: string;
  amount: Uint128;
}
export interface InstantiateMarketingInfo {
  description?: string | null;
  logo?: Logo | null;
  marketing?: string | null;
  project?: string | null;
}
// ...
export type TokenInfo = {
  existing: {
    address: string;
    staking_contract: StakingInfo;
  };
} | {
  new: {
    code_id: number;
    decimals: number;
    initial_balances: Cw20Coin[];
    initial_dao_balance?: Uint128 | null;
    label: string;
    marketing?: InstantiateMarketingInfo | null;
    name: string;
    staking_code_id: number;
    symbol: string;
    unstaking_duration?: Duration | null;
  };
};